ADD: SamjFill.LabelPresenceIndicatorAtGlobalCoord and installs the...
...samj.installPerSlicesTrackingPromptBehaviour()
and installRepeatPromptOnNextSliceBehaviour().

The class' indicator isPresent() may fast fail when
no 'label' is selected. Which is, for example, the case
when this code is installed as Labkit is yet not fully initiated
and cannot provide model.labeling(). On the other hand,
even when user doesn't click anything/doesn't choose any from
the labels, the first one is returned with model.selectedLabel()
and so it seems that there will always be some label available
anytime later during the operation of the application.